plotting function refactor
